requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
python-nmap>=0.7.1
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
scapy>=2.5.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0 